requirements.txt
python-dotenv==1.0.0
pandas==2.1.4
bardapi==0.1.38
sydney-py==0.17.6
markdown==3.5.1
nltk==3.6.2
ConTexto==0.2.0
matplotlib==3.8.2
seaborn==0.13.0
